requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.4
aiosignal==1.3.1
anyascii==0.3.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
bangla==0.0.2
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
boltons==23.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.6
clip-anytorch==2.5.2
colorama==0.4.6
comtypes==1.2.0
contourpy==1.1.0
coqpit==0.0.17
cycler==0.11.0
Cython==0.29.30
dateparser==1.1.8
decorator==4.4.2
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
encodec==0.1.1
filelock==3.12.2
Flask==2.3.2
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
g2pkk==0.1.2
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
inflect==5.6.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.1
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
k-diffusion==0.0.15
kiwisolver==1.4.4
kornia==0.6.12
lazy_loader==0.3
librosa==0.10.0.post2
llvmlite==0.40.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mecab==0.996.3
mecab-python3==1.0.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.0
pipreqs==0.4.13
pooch==1.6.0
proglog==0.1.10
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pynndescent==0.5.10
pyparsing==3.0.9
pypinyin==0.49.0
pypiwin32==223
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.5
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.7.18
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.15.2
tqdm==4.65.0
trainer==0.0.28
trampoline==0.1.2
transformers==4.31.0
TTS==0.15.6
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3==1.26.16
wandb==0.15.7
wcwidth==0.2.6
Werkzeug==2.3.6
yarg==0.1.9
yarl==1.9.2
